feat : RecommendPage.vue에 추천 기능 연동
